cache_app_demo/requirements.txt
Django==4.2.18
djangorestframework==3.14.0
google-cloud-pubsub>=2.5
python-dotenv==1.0.0
gunicorn==21.2.0
